Separate AltibaseLimitHandler
